requirements.txt
openai>=1.0.0
langchain>=0.1.0
streamlit>=1.30.0
requests>=2.28.0
python-dotenv>=1.0.0
tiktoken>=0.5.1
serpapi>=0.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyPDF2
streamlit-extras
